hm__6/pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["dima cheban <dimacheban23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.10"
faker = "^30.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
